+++ redaxo/src/addons/install/CHANGELOG.md
@@ -6,4 +6,5 @@ Version 2.0.1 – XX.XX.2016
 
 ### Bugfixes
 
+* Beim Download/Update großer Addons kam es teilweise zu einer Fehlermeldung
 * Api-Login und Api-Key werden nicht mehr in rex_config (in der DB) gespeichert, und sind somit nicht mehr im Export enthalten